# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Integration Runtime",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/file-grants": {
      "post": {
        "tags": [
          "T02"
        ],
        "summary": "Register Grant",
        "operationId": "registerFileGrant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_GrantPayload_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "post": {
        "tags": [
          "T02"
        ],
        "summary": "Create",
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_CreatePayload_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/project-sessions": {
      "post": {
        "tags": [
          "T02"
        ],
        "summary": "Open Project",
        "operationId": "openProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Session_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}": {
      "get": {
        "tags": [
          "T02"
        ],
        "summary": "Get Project",
        "operationId": "getProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Project_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/drafts/{draftId}": {
      "put": {
        "tags": [
          "T02"
        ],
        "summary": "Save Draft",
        "operationId": "saveDraft",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_SaveDraftPayload_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "T02"
        ],
        "summary": "Get Draft",
        "operationId": "getDraft",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Draft_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/drafts": {
      "get": {
        "tags": [
          "T02"
        ],
        "summary": "List Drafts",
        "operationId": "listDrafts",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_DraftSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/operations/{operationId}": {
      "get": {
        "tags": [
          "T02"
        ],
        "summary": "Project Operation",
        "operationId": "getProjectOperation",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Operation_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/{operationId}": {
      "get": {
        "tags": [
          "T02"
        ],
        "summary": "Operation",
        "operationId": "getGlobalOperation",
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Operation_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/private/recent-projects/{project_id}/directory": {
      "get": {
        "tags": [
          "T02"
        ],
        "summary": "Directory",
        "operationId": "getRecentProjectDirectory",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PrivateDirectory_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/project-sessions/{session_id}": {
      "delete": {
        "tags": [
          "T02"
        ],
        "summary": "Close",
        "operationId": "closeProjectSession",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ClosedSession_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Command_CreatePayload_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/CreatePayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[CreatePayload]"
      },
...
      "Command_GrantPayload_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/GrantPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[GrantPayload]"
      },
...
      "Command_SaveDraftPayload_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/SaveDraftPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[SaveDraftPayload]"
      },
...
      "CreatePayload": {
        "properties": {
          "directoryGrantId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "aspect": {
            "type": "string",
            "enum": [
              "16:9",
              "9:16"
            ],
            "title": "Aspect"
          },
          "resolution": {
            "type": "string",
            "enum": [
              "720p",
              "1080p"
            ],
            "title": "Resolution"
          },
          "fps": {
            "$ref": "#/components/schemas/Fps"
          },
          "targetMs": {
            "$ref": "#/components/schemas/PositiveInteger"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "directoryGrantId",
          "name",
          "aspect",
          "resolution",
          "fps",
          "targetMs"
        ],
        "title": "CreatePayload"
      },
...
      "Draft": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "artifactId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "baseRevisionId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "$ref": "#/components/schemas/DraftContent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "artifactId",
          "baseRevisionId",
          "content"
        ],
        "title": "Draft"
      },
...
      "DraftSummary": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "artifactId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/DraftKind"
          },
          "savedAt": {
            "type": "string",
            "title": "Savedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "artifactId",
          "kind",
          "savedAt"
        ],
        "title": "DraftSummary"
      },
...
      "Envelope_ClosedSession_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/ClosedSession"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[ClosedSession]"
      },
...
      "Envelope_Draft_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Draft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Draft]"
      },
...
      "Envelope_MutationReceipt_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/MutationReceipt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[MutationReceipt]"
      },
      "Envelope_Operation_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Operation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Operation]"
      },
      "Envelope_PrivateDirectory_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/PrivateDirectory"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[PrivateDirectory]"
      },
      "Envelope_Project_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Project"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Project]"
      },
...
      "Envelope_Session_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Session]"
      },
...
      "Envelope_list_DraftSummary__": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DraftSummary"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[list[DraftSummary]]"
      },
...
      "GrantPayload": {
        "properties": {
          "grantId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "path": {
            "type": "string",
            "maxLength": 32767,
            "minLength": 1,
            "title": "Path"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "createProject",
              "openProject",
              "importMedia",
              "ffmpeg"
            ],
            "title": "Purpose"
          },
          "windowId": {
            "$ref": "#/components/schemas/PositiveInteger"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "grantId",
          "path",
          "purpose",
          "windowId"
        ],
        "title": "GrantPayload"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MutationReceipt": {
        "properties": {
          "operationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "committedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "resourceId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "state": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operationId",
          "committedRevision",
          "resourceId",
          "state"
        ],
        "title": "MutationReceipt"
      },
...
      "OpenProject": {
        "properties": {
          "directoryGrantId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "requestedMode": {
            "type": "string",
            "enum": [
              "read",
              "write"
            ],
            "title": "Requestedmode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "directoryGrantId",
          "requestedMode"
        ],
        "title": "OpenProject"
      },
...
      "Project": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "revision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "eventSequence": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "formatVersion": {
            "$ref": "#/components/schemas/PositiveInteger"
          },
          "aspect": {
            "type": "string",
            "enum": [
              "16:9",
              "9:16"
            ],
            "title": "Aspect"
          },
          "resolution": {
            "type": "string",
            "enum": [
              "720p",
              "1080p"
            ],
            "title": "Resolution"
          },
          "fps": {
            "$ref": "#/components/schemas/Fps"
          },
          "targetMs": {
            "$ref": "#/components/schemas/PositiveInteger"
          },
          "budgetMicroCny": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "savedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Savedat"
          },
          "readOnly": {
            "type": "boolean",
            "title": "Readonly"
          },
          "executionMode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "synthetic",
                  "real"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Executionmode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "revision",
          "eventSequence",
          "formatVersion",
          "aspect",
          "resolution",
          "fps",
          "targetMs",
          "budgetMicroCny",
          "savedAt",
          "readOnly"
        ],
        "title": "Project"
      },
...
      "SafeInteger": {
        "type": "integer",
        "ge": 0,
        "le": 9007199254740991
      },
...
      "SaveDraftPayload": {
        "properties": {
          "draftId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "artifactId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "baseRevisionId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "$ref": "#/components/schemas/DraftContent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draftId",
          "artifactId",
          "baseRevisionId",
          "content"
        ],
        "title": "SaveDraftPayload"
      },
...
      "Session": {
        "properties": {
          "projectId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "projectSessionId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "mode": {
            "type": "string",
            "enum": [
              "read",
              "write"
            ],
            "title": "Mode"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "projectId",
          "projectSessionId",
          "mode",
          "project"
        ],
        "title": "Session"
      },
...
      "Uuid": {
        "type": "string",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "format": "uuid"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "RuntimeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "32-byte random token, base64url without padding"
      }
    }
  },
  "security": [
    {
      "RuntimeBearer": []
    }
  ]
}